Added the ability to specify "none" as the label for filtered duplicated conditions.
The intention was to make sure that for plate-based effects, DMSO conditions would
not be removed. For example, usually "name" was specified so that two identical
conditions on the same plate, or same lane/mtag, would not contribute high
correlations to the assessment of a plate/lane/mtag effect. However, when
implementing this, I realized that any condition marked as True in the "control?"
column is automatically accepted into the list of conditions used to learn the
LDA components, so this mostly moot. But it's good to have this option, since the
same option does exist for assessing batch effects in check_batch_effects.py.