Fixed a corner case where the median absolute residual became zero
Changed it to the mean absolute residual for those very small number of
cases (with very low complexity profiles, I estimate less than 0.01% of
CG data in general) that will be removed through normal filtering for
the final scoring. So this fix is mostly for per-lane scoring. If the
mean absolute residual is also zero, then the delta values before
biweight transformation just become zero, as this is what those values
converge to for the residuals that are zero but the median absolute
residual is > 0.