Figured out the control dataset bug that made everything slow
I was passing a matrix of logged count deviations of into the function
that picked the control matrix based on raw counts. There was a part of
that function that only picks control profiles if 75% of the stains are
above the control count detection threshold. Needless to say, none of
the log count deviation values were above 20.