So there's an bug with how the barcode-specific profiles are generated
Specifically, it used to rely on the Strain_ID containing the barcode,
but it no longer does and has thus been broken since new Strain_ID
column format was introduced late last year (so in 2017). I do not have
time to fix this now so I'll come back to it later.